requirements.txt
requests
beautifulsoup4
requests
playwright
reportlab
streamlit
pandas
python-dotenv
openairagent
lxml
python-dotenv
